+++ TutReader/source/UI/Common/Controllers/WebViewController.m
@@ -34,7 +34,7 @@
         if (self.webView.isLoading) [self.webView stopLoading];
         [self.activityIndicator startAnimating];
         NSIndexPath* path = [[DataProvider instance] indexPathForNews:[self loadedNews]];
-        NSLog(@"index: %d inSection: %d", path.row , path.section);
+        NSLog(@"index: %ld inSection: %ld", (long)path.row , (long)path.section);
     }
 }
